# Rename C overloads based on the typenames of their arguments