Previously, the following C code:

```
extern float Common(int*, int y);
extern float Common(int*, bool y);
extern float Common(int*, const char *y);
```

would generate bindings similar to:

```
extern "C" {
    #[link_name = "\u{1}_Z6CommonPii"]
    pub fn Common(arg1: *mut c_int, y: c_int) -> f32;
}
extern "C" {
    #[link_name = "\u{1}_Z6CommonPib"]
    pub fn Common1(arg1: *mut c_int, y: bool) -> f32;
}
extern "C" {
    #[link_name = "\u{1}_Z6CommonPiPKc"]
    pub fn Common2(arg1: *mut c_int, y: *const c_char) -> f32;
}
```

This patch changes the function suffixes to be based on each overload's
signature instead of just its order in the source file:

```
extern "C" {
    #[link_name = "\u{1}_Z6CommonPii"]
    pub fn Common(arg1: *mut c_int, y: c_int) -> f32;
}
extern "C" {
    #[link_name = "\u{1}_Z6CommonPib"]
    pub fn Common_bool(arg1: *mut c_int, y: bool) -> f32;
}
extern "C" {
    #[link_name = "\u{1}_Z6CommonPiPKc"]
    pub fn Common_ptr_const_char(arg1: *mut c_int, y: *const c_char) -> f32;
}
```

When does renaming occur:

 - If a function has no overloads, its Rust name is unchanged.
 - If there are overloads, the function which occurs first in the c file
   is unchanged in name (if possible*). The other overloads are renamed.

How is the new function name chosen for some particular overload:

 1. Strip out as many arguments from the left which are common to all overloads.
 2. Convert the remaining argument typenames to a string and join them with '_'.
 3. Append this to the original C function name along with another '_'.

How are argument types converted to string:
I use the `Type::canonical_type` method for this, which seems to follow these rules:

 - Types are converted based on their C name, with spaces replaced by '_'.
       unsigned int => 'unsigned_int'
       struct foo_bar => 'struct_foo_bar'
 - Pointers follow the above rules but prefixed with 'ptr_', or 'ptr_const_' for pointers to const values.
       struct foo_bar * => 'ptr_struct_foo_bar'
       const int * => 'ptr_const_int'
 - Typedefs are resolved to their underlying type before being converted.

Hopefully the updated test code serve to provide concrete examples of this renaming logic.

------

This renaming logic is what I'm finding useful in my practice. Actually, I'm thinking about adding logic to map `const char*` to `cstr` instead of `ptr_const_char`, and similarly mapping `unsigned int` to `uint` instead, removing `struct_` from the typenames, etc. The decision to not rename the first overload is similarly arbitrary (although it does match the previous behavior). It's also subject to an edge case (*) where an overload with zero arguments comes after an overload with arguments, in which case the zero-argument overload gets the canonical name and the first overload actually *is* renamed, to avoid collision.

So I'm not sure if this belongs behind some configuration flag, or if it requires more thought about how to enable user configuration first (some new fn on the `ParseCallbacks` trait?).

Thanks for your work in maintaining this incredibly useful project!