@Wallacoloo 2019-06-13
Grabbed some timings. But I won't likely have a chance to pursue handling templates over the next week.

After uncommenting the `panic!` line in tests/stylo_sanity.rs and enabling the `testing_only_libclang_4` feature (although the test seemed to work without it too -- I'm not sure why it was gated), I ran
```
$ cargo test --release sanity_check_can_generate_stylo_bindings
```

rustc version is 1.35.0 (stable).

Before this change (i.e. measuring against Wallacoloo:master -- not rust-lang:master) I got these numbers for "Generated Stylo bindings in: $y(s)":
```
5.167769976
5.128981535
4.664460779
5.14214074
4.779876099
```

After this change (i.e. measuring against Wallacoloo:topics/signature-based-overload-naming) (s):
```
5.28417303
4.474109496
5.137210242
4.947957164
5.22608733
```

There doesn't seem to be much of a perf difference... I'm half-convinced I somehow messed up the testing because of that. I tried again with a `cargo clean --release`d between each build to be defensive, but got the same numbers.

Full test output, in case anything looks fishy (before change):
```
running 1 test
test sanity_check_can_generate_stylo_bindings ...   time:   437.265 ms.	parse
  time:     0.000 ms.	process_replacements
  time:    14.003 ms.	deanonymize_fields
  time:   292.512 ms.	compute_whitelisted_and_codegen_items
  time:    35.260 ms.	compute_has_vtable
  time:   140.975 ms.	compute_sizedness
  time:    28.611 ms.	compute_has_destructor
  time:   116.223 ms.	find_used_template_parameters
  time:   162.493 ms.	compute_cannot_derive_debug
  time:     0.000 ms.	compute_cannot_derive_default
  time:   220.262 ms.	compute_cannot_derive_copy
  time:    30.614 ms.	compute_has_type_param_in_array
  time:     0.000 ms.	compute_has_float
  time:     0.000 ms.	compute_cannot_derive_hash
  time:     0.000 ms.	compute_cannot_derive_partialord_partialeq_or_eq
  time:   235.901 ms.	codegen
FAILED

failures:

---- sanity_check_can_generate_stylo_bindings stdout ----
/home/colin/src/rust-bindgen/tests/stylo.hpp:7505:32: warning: replacement function 'operator new' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7512:52: warning: replacement function 'operator new' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7518:52: warning: replacement function 'operator new[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7524:52: warning: replacement function 'operator new[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7530:52: warning: replacement function 'operator delete' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7536:52: warning: replacement function 'operator delete' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7542:52: warning: replacement function 'operator delete[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7548:52: warning: replacement function 'operator delete[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:123367:16: warning: offset of on non-standard-layout type 'js::BaseProxyHandler' [-Winvalid-offsetof], err: false


Generated Stylo bindings in: 4.769149313s


thread 'main' panicked at 'Uncomment this line to get timing logs', tests/stylo_sanity.rs:554:5
note: Run with `RUST_BACKTRACE=1` environment variable to display a backtrace.


failures:
    sanity_check_can_generate_stylo_bindings

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 0 filtered out
```

After change (I added a timer to compute_overload_suffixes):
```
running 1 test
test sanity_check_can_generate_stylo_bindings ...   time:   453.178 ms.	parse
  time:     0.001 ms.	process_replacements
  time:    18.440 ms.	deanonymize_fields
  time:   348.118 ms.	compute_whitelisted_and_codegen_items
  time:    37.587 ms.	compute_has_vtable
  time:   143.170 ms.	compute_sizedness
  time:    42.478 ms.	compute_has_destructor
  time:   163.069 ms.	find_used_template_parameters
  time:   187.824 ms.	compute_cannot_derive_debug
  time:     0.000 ms.	compute_cannot_derive_default
  time:   217.009 ms.	compute_cannot_derive_copy
  time:    39.252 ms.	compute_has_type_param_in_array
  time:     0.000 ms.	compute_has_float
  time:     0.000 ms.	compute_cannot_derive_hash
  time:     0.001 ms.	compute_cannot_derive_partialord_partialeq_or_eq
  time:    30.403 ms.	compute_overload_suffixes
  time:   226.141 ms.	codegen
FAILED

failures:

---- sanity_check_can_generate_stylo_bindings stdout ----
/home/colin/src/rust-bindgen/tests/stylo.hpp:7505:32: warning: replacement function 'operator new' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7512:52: warning: replacement function 'operator new' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7518:52: warning: replacement function 'operator new[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7524:52: warning: replacement function 'operator new[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7530:52: warning: replacement function 'operator delete' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7536:52: warning: replacement function 'operator delete' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7542:52: warning: replacement function 'operator delete[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:7548:52: warning: replacement function 'operator delete[]' cannot be declared 'inline' [-Winline-new-delete], err: false
/home/colin/src/rust-bindgen/tests/stylo.hpp:123367:16: warning: offset of on non-standard-layout type 'js::BaseProxyHandler' [-Winvalid-offsetof], err: false


Generated Stylo bindings in: 5.08529371s


thread 'main' panicked at 'Uncomment this line to get timing logs', tests/stylo_sanity.rs:554:5
note: Run with `RUST_BACKTRACE=1` environment variable to display a backtrace.


failures:
    sanity_check_can_generate_stylo_bindings

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 0 filtered out
```

Seems to consistently take around 30 ms to compute the overloads. I wouldn't expect it to affect the other sections, except maybe codegen, but if it did that was mostly lost in the noise.